:arrow_up: upgraded dp3t-sdk dependency to 0.1.7

+++ android/src/main/java/com/dpppt/DppptModule.java
@@ -59,7 +59,7 @@ public class DppptModule extends ReactContextBaseJavaModule {
         map.putInt("numberOfContacts", tracingStatus.getNumberOfContacts());
         map.putBoolean("isAdvertising", tracingStatus.isAdvertising());
         map.putBoolean("isReceiving", tracingStatus.isReceiving());
-        map.putLong("getLastSyncDate", tracingStatus.getLastSyncDate());
+        map.putInt("getLastSyncDate", tracingStatus.getLastSyncDate());
         promise.resolve(map);
     }
